pyproject.toml
[tool.poetry]
name = "nonebot_plugin_memes_api"
version = "0.3.0"
description = "Nonebot2 plugin for making memes"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-memes-api"
repository = "https://github.com/noneplugin/nonebot-plugin-memes-api"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-plugin-alconna = ">=0.37.0,<1.0.0"
nonebot-plugin-session = "^0.3.0"
nonebot-plugin-userinfo = "^0.2.0"
nonebot-plugin-localstore = "^0.6.0"
httpx = ">=0.20.0,<1.0.0"
pypinyin = ">=0.44.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
nonebot-adapter-onebot = "^2.4.0"

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
